LayoutTests/fast/viewport/ios/shrink-to-fit-content-responsive-viewport-with-horizontal-overflow-expected.txt

This test verifies that the shrink-to-fit-content heuristic prevents horizontal scrolling by shrinking the page, even when a page specifies a responsive viewport. To run the test manually, load the page and check that the bar entirely fits within the viewport, and the page is not horizontally scrollable.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS minScale is expectedScale
PASS innerWidth became 960
PASS successfullyParsed is true

TEST COMPLETE
